Add a function used to check if a gem is available for the given variant
This function can be useful when a gem can use another gem to supercharge its feature, without having that gem as a mandatory dependency.

Simple usage:
```cmake
ly_is_gem_available(
    NAME EMotionFX
    RETURN_VARIABLE result
    VALUE_IF_FOUND EMOTIONFX_ENABLED
    VALUE_IF_NOT_FOUND EMOTIONFX_DISABLED
    VARIANT Clients
)
```
With that we can now for example use the result variable in the compile definitions of the gem, so with EMOTIONFX_ENABLED define will enable some part of the gem code optimized for the use with the EMotionFX gem.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>