Switch RenderViewportWidget to tick with the renderer
This is an attempt to rectify some stuttering behavior in the camera interpolation of the modular camera controller when ticking with the tick bus instead.

Note: In poking around to look at this, I did notice that for DX12, we seem to synchronously wait for VSync in CommandQueueContext::End. This means that we're not capturing valid frame timings anywhere, and I believe it also means we're presently starving the system tick bus, as render tick is on the main thread at the moment. This may warrant further investigation.

Signed-off-by: nvsickle <nvsickle@amazon.com>