SynergyInput Gem.
Synergy input was lost when we removed the CryLegacy Gem. This resurrects the code and updates it to be more conformant with the AzFramework Input framework.

The majority of the code in SynergyInputClient.cpp is still entirely unmodified from the original Cry drop and could use some love, but it used to work at least.

I have unfortunately been unable to test this, but if anyone ever needs Synergy integration this should get us ~95% of the way there.

Signed-off-by: bosnichd <bosnichd@amazon.com>