[MeshOptimizer] Return the correct size from the position map when there are blendshapes (#2391)
The previous version returned the incorrect size from the position map when
blendshapes are present in the model. When there are blendshapes, the
vertex welding is disabled, and nothing is inserted into the position map.
However, the position map's size was being used to dictate how many
elements to create in the skinning info. The skinning info tries to
compensate for an incorrect max vertex index by resizing its underlying
vector when adding an influence, but that was using the index as the new
size of the vector, so it was off by one. This fixes both errors.

Signed-off-by: Chris Burel <burelc@amazon.com>
Signed-off-by: Benjamin Jillich <jillich@amazon.com>