Don't require a specific version of Wwise SDK, allow for later versions to be used (#2484)
* Updates the Wwise SDK detection in CMake

Parses AkWwiseSDKVersion.h to obtain the version instead of requiring an
exact version to be installed.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>

* Updates readme.md information about Wwise

Rewrite some of the setup info about Wwise, mention the two ways in
which Wwise can be found via CMake and that users can install later
versions than the one specified.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>

* Remove unnecessary parts of Wwise version detect

Don't need to parse the file contents, it was only being used to print a
message to CMake.  Simplify the validation to just checking that the SDK
version header exists.  Maybe in the future we can reinstate better
version checking to make sure the version falls within a range.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>

* Updates capitalization of Audiokinetic name

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>