Fix a crash when reloading AudioControlEditor controls (#2729)
* Fix a crash when reloading ACE controls data

The crash was due to destruction of xml_node that was held in a
unique_ptr.  Rapidxml has a very rudimentary memory allocation design,
so in most cases dynamic allocations aren't even made.  The memory_pool
does all the cleanup in its destructor, so having a unique_ptr run its
default_delete was causing the crash.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>

* Fix numerical conversion warnings

Wwise source files needed a few fixes for the numerical conversion
warning changes that went in recently.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>