EMotion FX: Asset Processor generates an Actor for every FBX (#4284)
Fixing the actor behavior to ignore actor asset generation for files that only contain nodes/bones. We're only checking for blend shapes and skins now. In case neither is present in an FBX, we don't export the actor asset.

This makes a plain skeleton a special case where users manually need to add the actor group for the fbx. The most common use case for these are animation files anyway, where we don't want to export actor assets.

Also increased the version number of the actor group so that they are all getting reprocessed and the unneeded actor assets get removed.

Signed-off-by: Benjamin Jillich <jillich@amazon.com>