Created feature matrix and moved all feature data into it
We now have a new FeatureMatrix which is responsible for storing the feature data that we need for the motion matching algorithm in a cache-efficient way.
The great thing is that this is also the enabler for any feature analysis and later on machine learning as we're in the right data format already.

The FeatureMatrix internally stores the data in a 2D dense matrix from the Eigen library. This can easily be replaced with another linear algebra/vector/matrix library though and the FeatureMatrix acts as a wrapper.

We're currently extracting the following motion features from our motion database:
Left foot position/velocity, right foot position/velocity and the root joint trajectory with 6 past and 6 future sample positions and directions.

Having 41203 keyframes/poses in our motion database, this results in a feature matrix holding 22.63 MB of data. This is only the data size for the extracted features.

Signed-off-by: Benjamin Jillich <jillich@amazon.com>