EMotion FX: Adding/Removing/Changing Components to an Entity with Actor Component erases the "Motion extraction joint"
Applying meta data onto actors failed as the commands use the actor manager to search them by ID but the actor manager now stores actor assets. The temporarily created actor in the actor exporter was not registered at the actor manager for that reason and the commands were failing to find the actor. This resulted in that all meta data for actors could not be applied on the actors.

Signed-off-by: Benjamin Jillich <jillich@amazon.com>